+++ src/DevBoost.DroneDelivery.Domain/Entities/Cliente.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Text;
 
 namespace DevBoost.DroneDelivery.Domain.Entities
 {
@@ -16,7 +14,7 @@ namespace DevBoost.DroneDelivery.Domain.Entities
 
         public Guid Id { get; set; }
         public string Nome { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
     }
 }
